♻️ use the correct metadata for the search page
