+++ apps/web/app/(entity)/[network]/[...path]/page.tsx
@@ -8,49 +8,13 @@ import {
   UnhealthyNetworkError,
 } from "~/lib/headless-utils";
 import { Table } from "~/ui/entity/table";
-import { capitalize, parseHeadlessRouteVercelFix } from "~/lib/shared-utils";
+import { getMetadata, parseHeadlessRouteVercelFix } from "~/lib/shared-utils";
 import { notFound } from "next/navigation";
 import { getSingleNetworkCached } from "~/lib/network";
 import { displayFiltersSchema } from "~/lib/display-filters";
 import { ALWAYS_ONLINE_NETWORKS } from "~/lib/constants";
 import { ShallowPush } from "~/ui/shallow-push";
-
-function shortenId(str: string) {
-  if (str.length < 12) {
-    return str;
-  }
-  if (str.match(/^(?:[0-9]+\.){3}[0-9]+$/)) {
-    return str.substring(0, 12) + "...";
-  }
-  return str.slice(0, 6) + "..." + str.slice(-6);
-}
-
-function formatTypeName(str: string, singular?: boolean) {
-  const pluralToSingular: Record<string, string> = {
-    addresses: "address",
-  };
-
-  const formatMap: Record<string, string> = {
-    eth: "ETH",
-    spl: "SPL",
-  };
-  const parts = str.split("-");
-
-  return parts
-    .map((part, index) => {
-      if (singular && index === parts.length - 1) {
-        let mappedSingular = pluralToSingular[part.toLowerCase()];
-        if (mappedSingular) {
-          return capitalize(mappedSingular);
-        }
-        if (part.endsWith("s")) {
-          return capitalize(str.slice(0, -1));
-        }
-      }
-      return formatMap[part.toLowerCase()] || capitalize(part);
-    })
-    .join(" ");
-}
+import { SearchMetadata } from "~/ui/search-metadata";
 
 export async function generateMetadata({
   params: _params,
@@ -58,56 +22,20 @@ export async function generateMetadata({
   params: HeadlessRoute;
 }) {
   const params = parseHeadlessRouteVercelFix(_params);
+
+  const network = await getSingleNetworkCached(params.network);
+  if (!network) {
+    return {};
+  }
+
   if (params.path[0] === "search") {
     const query = params.path[1];
     return {
       title: `Searching for ${query}`,
     };
   }
-  const network = await getSingleNetworkCached(params.network);
-  if (!network) {
-    return {};
-  }
 
-  // Special cases
-  if (
-    params.path.length === 1 &&
-    (params.path[0] === "transactions" || params.path[0] === "blocks")
-  ) {
-    return {
-      title: `Latest ${capitalize(params.path[0])} on ${capitalize(
-        network.brand,
-      )} ${capitalize(network.chainName)}`,
-      description: `Latest ${capitalize(params.path[0])} on ${capitalize(
-        network.brand,
-      )} ${capitalize(network.chainName)}, brought to you by Modular Cloud.`,
-    };
-  }
-
-  let titleParts = [];
-  let descriptionParts = [];
-  for (let i = 0; i < params.path.length; i += 2) {
-    const type = params.path[i];
-    const id = params.path[i + 1];
-
-    if (!id) {
-      titleParts.unshift(formatTypeName(type));
-      descriptionParts.unshift(formatTypeName(type));
-      break;
-    }
-
-    titleParts.unshift(`${formatTypeName(type, true)} ${shortenId(id)}`);
-    descriptionParts.unshift(`${formatTypeName(type, true)} ${id}`);
-  }
-
-  return {
-    title: `${titleParts.join(" - ")} - ${capitalize(
-      network.brand,
-    )} ${capitalize(network.chainName)}`,
-    description: `${descriptionParts.join(" in ")} on ${capitalize(
-      network.brand,
-    )} ${capitalize(network.chainName)}, brought to you by Modular Cloud.`,
-  };
+  return getMetadata(params, network);
 }
 
 export default function EntityPage({
@@ -133,23 +61,21 @@ async function AyncPageContent({
   params: HeadlessRoute;
   searchParams?: Record<string, any>;
 }) {
+  const network = await getSingleNetworkCached(params.network);
+  if (!network) {
+    notFound();
+  }
+
   const entityType = params.path[0];
 
   let redirectTo: string | null = null;
 
   if (entityType === "search") {
     const query = params.path[1];
-    const [networkResult, searchResult, networkStatusResult] =
-      await Promise.allSettled([
-        getSingleNetworkCached(params.network),
-        search(params.network, query),
-        checkIfNetworkIsOnline(params.network),
-      ]);
-
-    const network = networkResult.status === "fulfilled" && networkResult.value;
-    if (!network) {
-      notFound();
-    }
+    const [searchResult, networkStatusResult] = await Promise.allSettled([
+      search(params.network, query),
+      checkIfNetworkIsOnline(params.network),
+    ]);
 
     if (
       !ALWAYS_ONLINE_NETWORKS.includes(network.brand) &&
@@ -173,7 +99,19 @@ async function AyncPageContent({
   if (page.body.type === "notebook") {
     return (
       <>
-        {redirectTo && <ShallowPush path={redirectTo} />}
+        {redirectTo && (
+          <>
+            <ShallowPush path={redirectTo} />
+            <SearchMetadata
+              correctPath={redirectTo}
+              network={{
+                brand: network.brand,
+                chainName: network.chainName,
+                slug: network.slug,
+              }}
+            />
+          </>
+        )}
         <Overview properties={page.body.properties} isIBC={page.isIBC} />
       </>
     );
@@ -181,7 +119,19 @@ async function AyncPageContent({
 
   return (
     <>
-      {redirectTo && <ShallowPush path={redirectTo} />}
+      {redirectTo && (
+        <>
+          <ShallowPush path={redirectTo} />
+          <SearchMetadata
+            correctPath={redirectTo}
+            network={{
+              brand: network.brand,
+              chainName: network.chainName,
+              slug: network.slug,
+            }}
+          />
+        </>
+      )}
       <Table initialData={page} route={params} />
     </>
   );
