+++ apps/web/lib/shared-utils.ts
@@ -194,3 +194,97 @@ export function arrayGroupByTo2DArray<T extends Record<string, any>>(
 
   return Array.from(map.values());
 }
+
+function shortenId(str: string) {
+  if (str.length < 12) {
+    return str;
+  }
+  if (str.match(/^(?:[0-9]+\.){3}[0-9]+$/)) {
+    return str.substring(0, 12) + "...";
+  }
+  return str.slice(0, 6) + "..." + str.slice(-6);
+}
+
+function formatTypeName(str: string, singular?: boolean) {
+  const pluralToSingular: Record<string, string> = {
+    addresses: "address",
+  };
+
+  const formatMap: Record<string, string> = {
+    eth: "ETH",
+    spl: "SPL",
+  };
+  const parts = str.split("-");
+
+  return parts
+    .map((part, index) => {
+      if (singular && index === parts.length - 1) {
+        let mappedSingular = pluralToSingular[part.toLowerCase()];
+        if (mappedSingular) {
+          return capitalize(mappedSingular);
+        }
+        if (part.endsWith("s")) {
+          return capitalize(str.slice(0, -1));
+        }
+      }
+      return formatMap[part.toLowerCase()] || capitalize(part);
+    })
+    .join(" ");
+}
+
+export function getMetadata(
+  routeParams: HeadlessRoute,
+  network: {
+    brand: string;
+    chainName: string;
+  },
+) {
+  const params = parseHeadlessRouteVercelFix(routeParams);
+
+  if (params.path[0] === "search") {
+    const query = params.path[1];
+    return {
+      title: `Searching for ${query}`,
+    };
+  }
+
+  // Special cases
+  if (
+    params.path.length === 1 &&
+    (params.path[0] === "transactions" || params.path[0] === "blocks")
+  ) {
+    return {
+      title: `Latest ${capitalize(params.path[0])} on ${capitalize(
+        network.brand,
+      )} ${capitalize(network.chainName)}`,
+      description: `Latest ${capitalize(params.path[0])} on ${capitalize(
+        network.brand,
+      )} ${capitalize(network.chainName)}, brought to you by Modular Cloud.`,
+    };
+  }
+
+  let titleParts = [];
+  let descriptionParts = [];
+  for (let i = 0; i < params.path.length; i += 2) {
+    const type = params.path[i];
+    const id = params.path[i + 1];
+
+    if (!id) {
+      titleParts.unshift(formatTypeName(type));
+      descriptionParts.unshift(formatTypeName(type));
+      break;
+    }
+
+    titleParts.unshift(`${formatTypeName(type, true)} ${shortenId(id)}`);
+    descriptionParts.unshift(`${formatTypeName(type, true)} ${id}`);
+  }
+
+  return {
+    title: `${titleParts.join(" - ")} - ${capitalize(
+      network.brand,
+    )} ${capitalize(network.chainName)}`,
+    description: `${descriptionParts.join(" in ")} on ${capitalize(
+      network.brand,
+    )} ${capitalize(network.chainName)}, brought to you by Modular Cloud.`,
+  };
+}
