+++ apps/web/ui/shallow-push/index.tsx
@@ -1,5 +1,5 @@
 "use client";
-import React from "react";
+import * as React from "react";
 
 interface Props {
   path: string;
